Entry point AddHTMLMetaAndTitle.php
CI requires a PHP entry point which is named alike the extension. That
has been done in master with f542afb.

Do a similar thing but instead of renaming the file, just introduce the
entry point with a single require_once.  That will prevent the file from
being overriden when people update their REL1_26 version.

Change-Id: Ic2073d81c42b150671486458f6f3b2154c1d1c4a